Clear the build warnings, and bring the docs up to the API
Both presets now build clean on the default preset and the player one.
Nothing here changes behaviour; the tests are the same 240 cases, 227
passed, 13 skipped, 0 failed.

What the warnings were hiding, in the two places it mattered:
note_renderer's Impl took its parameter under the same name as the member
initialised from it, so the ctor-init-list read the parameter and the
member order was wrong beside it -- -Wshadow and -Wreorder describing one
mistake from two angles. And bank_builder discarded a [[nodiscard]] name
from tone_zones, which is fine and now says so.

The rest is mechanical: casting both operands rather than one, so the
"cast to double" idiom stops promoting the other half implicitly; walking
spans with size_t; and renaming three inner bindings that shadowed outer
ones with the same meaning but a different scope -- the bank_builder pair
needed their uses moved with them, which is exactly the bug -Wshadow
exists to catch.

The docs had gone stale in three places against the event work:

- verification said the MIDI pipeline "is not implemented, and that is a
  decision". It is implemented, off by default, and the section now says
  which switch, what each stage is worth -- 128 for staging, 1 for the
  output stage -- and that 19 samples remain unattributed rather than
  assigned to whichever component is nearest.
- getting-started said events "land on the block boundary". They land at
  the top of the next render, or wherever `send_channel_at` places them.
- architecture listed only the two settings that let the engine exceed
  the module, and now lists the two that make it match the module's
  timing at a cost the engine does not otherwise carry.

Regenerated with `doxygen docs/Doxyfile`: every new symbol resolves and
no new warning appears.

+++ src/realtime/tone_generator.cpp
@@ -2081,8 +2081,9 @@ void ToneGenerator::Impl::mix_effects(std::span<float> left, std::span<float> ri
         }
         for (int n = 0; n < block_size; ++n) {
             const auto i = static_cast<std::size_t>(n);
-            left[i] += static_cast<float>(wet_left[i] * level_gains[i]);
-            right[i] += static_cast<float>(wet_right[i] * level_gains[i]);
+            left[i] += static_cast<float>(static_cast<double>(wet_left[i]) * level_gains[i]);
+            right[i] +=
+                static_cast<float>(static_cast<double>(wet_right[i]) * level_gains[i]);
         }
     };
